pyproject.toml
[project]
name = "coursework"
version = "0.1.0"
description = ""
authors = [
    {name = "vooterr",email = "zigl.zigl535@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10.0,<=3.13"
dependencies = [
    "transformers (>=4.57,<5.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "hf-xet (>=1.2.0,<2.0.0)",
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "torchaudio (>=2.5.1,<3.0.0)",
    "bitsandbytes (==0.43.1)",
    "accelerate (==0.30.1)",
    "sentencepiece (>=0.2.1,<0.3.0)",
    "protobuf (>=6.33.3,<7.0.0)",
    "tokenizers (>=0.22.2,<0.23.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-cu121"}
torchvision = {source = "pytorch-cu121"}
torchaudio = {source = "pytorch-cu121"}
